(maybe) fixed the ^ issue
The time display works with a nifty ASCII trick: it adds a constant to
each of the integer values for tens of seconds, seconds, and tenths of
seconds.  It does this in a for loop after the time is set.  In doing
so, it changed the character '.' into '^'.  I added a line to reset it
to '.' after the for loop.